Some data types changes

+++ first_example/drivers/Inc/stm32l476xx_gpio_driver.h
@@ -89,7 +89,7 @@ void GPIO_ToggleOutputPin(GPIO_RegDef_t* pGPIOx, uint8_t PinNumber);
 /*IRQ Configuration and Handling*/
 void GPIO_IRQITConfig(uint8_t IRQNumber, uint8_t EnorDi);
 void GPIO_Handling(uint8_t PinNumber);
-void GPIO_IRQPRConfig(uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_IRQPRConfig(uint8_t IRQNumber, uint32_t IRQPriority);
