+++ first_example/drivers/Src/stm32l476xx_gpio_driver.c
@@ -315,7 +315,7 @@ void GPIO_IRQITConfig(uint8_t IRQNumber, uint8_t EnorDi){
 
 	}
 }
-void GPIO_IRQPRConfig(uint8_t IRQNumber, uint8_t IRQPriority){
+void GPIO_IRQPRConfig(uint8_t IRQNumber, uint32_t IRQPriority){
 
 	uint8_t iprx=IRQNumber/4;
 	uint8_t iprx_section= IRQNumber%4;
